+++ packages/app/src/components/Root.tsx
@@ -3,13 +3,9 @@ import { UpsetConfig } from "@visdesignlab/upset2-core"
 import { Box, css } from "@mui/material"
 import { Body } from "./Body"
 import Header from "./Header"
-import { useRef, useState, useEffect, createContext, useMemo } from "react"
-import React from "react"
+import { useRef, useState, useEffect, useMemo } from "react"
 import Footer from "./Footer"
-import { useRecoilValue } from "recoil"
-import { queryParamAtom } from "../atoms/queryParamAtom"
 import { Home } from "./Home"
-import { getMultinetSession } from "../api/session"
 
 type Props = {
     provenance: UpsetProvenance,
@@ -17,11 +13,6 @@ type Props = {
     data: any,
     config?: UpsetConfig
 }
-
-export const ProvenanceContext = createContext<{
-    provenance: UpsetProvenance;
-    actions: UpsetActions;
-  }>(undefined!);
 
 export const Root = ({provenance, actions, data, config}: Props) => {
     const headerDiv = useRef<HTMLDivElement>(null);
@@ -44,46 +35,21 @@ export const Root = ({provenance, actions, data, config}: Props) => {
   
       setHeaderHeight(current.clientHeight);
     }, [headerHeight, headerDiv]);
-
-    const { workspace, sessionId } = useRecoilValue(queryParamAtom);
   
-    async function restoreSession() {
-      if (sessionId) {
-        const session = await getMultinetSession(workspace || '', sessionId);
-  
-        // Load the session if the object is not empty
-        if (typeof session.state === 'object' && Object.keys(session.state).length !== 0) {
-          provenance.importObject(session.state);
-        }
-      }
-    }
-  
-    useEffect(() => {
-      restoreSession();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
-
     return (
-        <ProvenanceContext.Provider
-            value={{
-                provenance,
-                actions,
-            }}
-        >
-            <div css={AppCss}>
-                <Box
-                    sx={{
-                        zIndex: theme => theme.zIndex.drawer + 1,
-                        position: 'relative',
-                    }}
-                    ref={headerDiv}
-                >
-                    <Header data={data}/>
-                </Box>
-                {data === null && <Home />}
-                <Body data={data} config={config}/>
-                <Footer />
-            </div>
-        </ProvenanceContext.Provider>
+      <div css={AppCss}>
+          <Box
+              sx={{
+                  zIndex: theme => theme.zIndex.drawer + 1,
+                  position: 'relative',
+              }}
+              ref={headerDiv}
+          >
+              <Header data={data}/>
+          </Box>
+          {data === null && <Home />}
+          <Body data={data} config={config}/>
+          <Footer />
+      </div>
     )
 }
